Extracted dummy data to new file
 - Extracted DummyData to new file
 - Added clarifying comments to MoltimateContainer and FilteredLigandResults
 - Removed commented code from LigandResultItem
